+++ test/libsolidity/ASTJSON/using_for_directive.sol
@@ -1,3 +1,6 @@
-library L {} contract C { using L for uint; }
+using {f} for uint;
+library L {}
+function f(uint) {}
+contract C { using L for *; }
 
 // ----