+++ test/libsolidity/syntaxTests/using/module_3.sol
@@ -1,0 +1,14 @@
+==== Source: A ====
+function id(uint x) pure returns (uint) {
+    return x;
+}
+
+==== Source: B ====
+import "A" as M;
+
+contract C {
+    using {M.id} for uint;
+    function f(uint x) public pure returns (uint) {
+        return x.id();
+    }
+}