+++ test/libsolidity/syntaxTests/using/using_lhs_asterisk_contract.sol
@@ -1,0 +1,5 @@
+contract C {
+    using * for uint;
+}
+// ----
+// ParserError 2314: (23-24): Expected identifier but got '*'