+++ test/libsolidity/syntaxTests/using/using_library_ast_file_level.sol
@@ -1,0 +1,4 @@
+library L { }
+using L for *;
+// ----
+// SyntaxError 8118: (14-28): The type has to be specified explicitly at file level (cannot use '*').